Split true-on-policy MoE helpers
